Add support for the 12-byte version of READ 28h
MMC-6 defines the 12-byte version alongside the 10-byte version. Both support
FUA, so it might be possible that some drives don't support FUA in the 10-byte
incarnation, but do in the 12-byte one, or only support the 12-byte version.

Unfortunately, this means that operation codes by themselves cannot be used as
command identifiers anymore, so this required some changes all around.